进一步完善

+++ script/build.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [[ ! -f main.go ]]; then
+    echo "invalid dir"
+    exit 1
+fi
+
 if [[ -f logdog ]]; then
     echo "logdog exists, remove"
     rm logdog
@@ -10,5 +15,7 @@ if [[ -f logdog.tar.gz ]]; then
     rm logdog.tar.gz
 fi
 
-go build -o logdog .
+go build -o logdog ./
 tar czf logdog.tar.gz logdog logdog.toml
+python3 ./script/upload.py
+rm logdog.tar.gz
